tasks/medical_chatbot/requirements.hidden.txt
pytest
python-dotenv
langchain-huggingface
langchain-community
langchain-groq
faiss-cpu
